fix(terminal): unwedge xterm after command completion — enable proposed API and contain OSC 133 handler errors
Every command completing in an integrated shell wedged the terminal: output
rendered fully, then no new prompt and no keypress echo, while the shell sat
healthy at a zle read. Root cause: registerDecoration is proposed API in
@xterm/xterm 6 and the Terminal never set allowProposedApi, so the OSC 133 'D'
handler threw inside xterm's parser. An exception escaping _innerWrite kills
xterm's write loop permanently — every subsequent chunk queues forever.

Two layers:
- xtermConfig.ts: shared XTERM_OPTIONS with allowProposedApi: true, restoring
  the shell-integration gutter markers (#47) that have thrown since ship.
- shellIntegration.ts: contain any error thrown by the OSC 133 handler so a
  decoration failure can never stall the parser again; warn once instead.

Verified live against a real PTY: burst output (seq 1 20000) now returns a
prompt, markers register, decoration elements render, write queue drains.
