+++ frontend/src/components/Terminal/shellIntegration.ts
@@ -63,7 +63,24 @@ export function createShellIntegration(
     return { dispose() {} };
   }
 
+  let warned = false;
+
   const handler = parser.registerOscHandler(133, (data: string): boolean => {
+    // An exception escaping an OSC handler kills xterm's write loop: the
+    // terminal permanently stops rendering output (it looks wedged, while the
+    // shell stays alive). Decorations are cosmetic — contain any failure here.
+    try {
+      handle(data);
+    } catch (err) {
+      if (!warned) {
+        warned = true;
+        console.warn('Shell integration marker failed; continuing without decorations.', err);
+      }
+    }
+    return true; // handled — xterm will not print the sequence
+  });
+
+  function handle(data: string): void {
     const parts = data.split(';');
     switch (parts[0]) {
       case 'A': {
@@ -96,8 +113,7 @@ export function createShellIntegration(
         }
         break;
     }
-    return true; // handled — xterm will not print the sequence
-  });
+  }
 
   function discard(block: PromptBlock): void {
     for (const d of block.decorations) d.dispose();
